Refactor stubing upgradeable contracts in CLI tests
