+++ packages/cli/src/test/models/Dependency.test.js
@@ -1,31 +1,20 @@
 'use strict';
 
-require('../setup');
+import { stubContractUpgradeable } from '../setup';
 
 import { expect } from 'chai';
 import { accounts } from '@openzeppelin/test-environment';
-
-import { Contracts } from '@openzeppelin/upgrades';
 
 import sinon from 'sinon';
 import npm from 'npm-programmatic';
 import Dependency from '../../models/dependency/Dependency';
 
+const sandbox = sinon.createSandbox();
+
 describe('Dependency', function() {
   const [from] = accounts;
 
-  beforeEach('stub getFromPathWithUpgradeable to simulate transpilation of contracts', async function() {
-    // stub getFromPathWithUpgradeable to fill upgradeable field with the same contract
-    sinon.stub(Contracts, 'getFromPathWithUpgradeable').callsFake(function(targetPath, contractName) {
-      const contract = Contracts.getFromPathWithUpgradeable.wrappedMethod.apply(this, [targetPath, contractName]);
-      contract.upgradeable = contract;
-      return contract;
-    });
-  });
-
-  afterEach(function() {
-    sinon.restore();
-  });
+  stubContractUpgradeable(sandbox);
 
   describe('static methods', function() {
     describe('#satisfiesVersion', function() {
@@ -84,13 +73,13 @@ describe('Dependency', function() {
 
     describe('#install', function() {
       it('calls npm install', async function() {
-        const npmInstallStub = sinon.stub(npm, 'install');
+        const npmInstallStub = sandbox.stub(npm, 'install');
         const nameAndVersion = 'mock-stdlib@1.1.0';
         const npmParams = { save: true, cwd: process.cwd() };
 
         await Dependency.installFn(nameAndVersion);
         npmInstallStub.should.have.been.calledWithExactly([nameAndVersion], npmParams);
-        sinon.restore();
+        sandbox.restore();
       });
     });
