+++ packages/cli/src/test/scripts/check.test.js
@@ -1,9 +1,8 @@
 'use strict';
-require('../setup');
+
+import { stubContractUpgradeable } from '../setup';
 
 import sinon from 'sinon';
-
-import { Contracts } from '@openzeppelin/upgrades';
 
 import CaptureLogs from '../helpers/captureLogs';
 import check from '../../scripts/check';
@@ -11,23 +10,14 @@ import ProjectFile from '../../models/files/ProjectFile';
 
 const expect = require('chai').expect;
 
+const sandbox = sinon.createSandbox();
+
 describe('check script', function() {
   beforeEach('setup', async function() {
     this.projectFile = new ProjectFile('mocks/packages/package-empty.zos.json');
   });
 
-  beforeEach('stub getFromPathWithUpgradeable to simulate transpilation of contracts', async function() {
-    // stub getFromPathWithUpgradeable to fill upgradeable field with the same contract
-    sinon.stub(Contracts, 'getFromPathWithUpgradeable').callsFake(function(targetPath, contractName) {
-      const contract = Contracts.getFromPathWithUpgradeable.wrappedMethod.apply(this, [targetPath, contractName]);
-      contract.upgradeable = contract;
-      return contract;
-    });
-  });
-
-  afterEach(function() {
-    sinon.restore();
-  });
+  stubContractUpgradeable(sandbox);
 
   beforeEach('capturing log output', function() {
     this.logs = new CaptureLogs();
