+++ packages/cli/src/test/scripts/create.test.js
@@ -1,5 +1,6 @@
 'use strict';
-require('../setup');
+
+import { stubContractUpgradeable } from '../setup';
 
 import { random } from 'lodash';
 import sinon from 'sinon';
@@ -16,9 +17,6 @@ import link from '../../scripts/link';
 import ProjectFile from '../../models/files/ProjectFile';
 import { ProxyType } from '../../scripts/interfaces';
 import NetworkFile from '../../models/files/NetworkFile';
-
-import * as Compiler from '../../models/compiler/Compiler';
-import * as transpiler from '../../transpiler';
 
 const should = require('chai').should();
 
@@ -40,16 +38,7 @@ describe('create script', function() {
   const version = '0.4.0';
   const txParams = { from: owner };
 
-  beforeEach('stub getFromPathWithUpgradeable to simulate transpilation of contracts', async function() {
-    // stub getFromPathWithUpgradeable to fill upgradeable field with the same contract
-    sandbox.stub(Contracts, 'getFromPathWithUpgradeable').callsFake(function(targetPath, contractName) {
-      const contract = Contracts.getFromPathWithUpgradeable.wrappedMethod.apply(this, [targetPath, contractName]);
-      contract.upgradeable = contract;
-      return contract;
-    });
-    sandbox.stub(Compiler, 'compile');
-    sandbox.stub(transpiler, 'transpileAndSave');
-  });
+  stubContractUpgradeable(sandbox);
 
   afterEach(function() {
     sandbox.restore();
