+++ packages/cli/src/test/scripts/freeze.test.js
@@ -1,18 +1,16 @@
 'use strict';
-require('../setup');
+
+import { stubContractUpgradeable } from '../setup';
 
 import sinon from 'sinon';
 
-import { Package, Contracts } from '@openzeppelin/upgrades';
+import { Package } from '@openzeppelin/upgrades';
 import { accounts } from '@openzeppelin/test-environment';
 
 import push from '../../scripts/push';
 import freeze from '../../scripts/freeze';
 import ProjectFile from '../../models/files/ProjectFile';
 import NetworkFile from '../../models/files/NetworkFile';
-
-import * as Compiler from '../../models/compiler/Compiler';
-import * as transpiler from '../../transpiler';
 
 const sandbox = sinon.createSandbox();
 
@@ -22,16 +20,7 @@ describe('freeze script', function() {
   const network = 'test';
   const txParams = { from: owner };
 
-  beforeEach('stub getFromPathWithUpgradeable to simulate transpilation of contracts', async function() {
-    // stub getFromPathWithUpgradeable to fill upgradeable field with the same contract
-    sandbox.stub(Contracts, 'getFromPathWithUpgradeable').callsFake(function(targetPath, contractName) {
-      const contract = Contracts.getFromPathWithUpgradeable.wrappedMethod.apply(this, [targetPath, contractName]);
-      contract.upgradeable = contract;
-      return contract;
-    });
-    sandbox.stub(Compiler, 'compile');
-    sandbox.stub(transpiler, 'transpileAndSave');
-  });
+  stubContractUpgradeable(sandbox);
 
   afterEach(function() {
     sandbox.restore();
