+++ packages/cli/src/test/scripts/publish.test.js
@@ -1,9 +1,10 @@
 'use strict';
-require('../setup');
+
+import { stubContractUpgradeable } from '../setup';
 
 import sinon from 'sinon';
 
-import { Contracts, App, Package, ImplementationDirectory, Proxy } from '@openzeppelin/upgrades';
+import { App, Package, ImplementationDirectory, Proxy } from '@openzeppelin/upgrades';
 import { accounts } from '@openzeppelin/test-environment';
 
 import publish from '../../scripts/publish';
@@ -11,9 +12,6 @@ import push from '../../scripts/push';
 import create from '../../scripts/create';
 import ProjectFile from '../../models/files/ProjectFile';
 import NetworkFile from '../../models/files/NetworkFile';
-
-import * as Compiler from '../../models/compiler/Compiler';
-import * as transpiler from '../../transpiler';
 
 const should = require('chai').should();
 
@@ -29,20 +27,7 @@ describe('publish script', function() {
   const contractName = 'ImplV1';
   const dependencyName = 'mock-stdlib-undeployed';
 
-  beforeEach('stub getFromPathWithUpgradeable to simulate transpilation of contracts', async function() {
-    // stub getFromPathWithUpgradeable to fill upgradeable field with the same contract
-    sandbox.stub(Contracts, 'getFromPathWithUpgradeable').callsFake(function(targetPath, contractName) {
-      const contract = Contracts.getFromPathWithUpgradeable.wrappedMethod.apply(this, [targetPath, contractName]);
-      contract.upgradeable = contract;
-      return contract;
-    });
-    sandbox.stub(Compiler, 'compile');
-    sandbox.stub(transpiler, 'transpileAndSave');
-  });
-
-  afterEach(function() {
-    sandbox.restore();
-  });
+  stubContractUpgradeable(sandbox);
 
   beforeEach('pushing package', async function() {
     const projectFile = new ProjectFile('mocks/packages/package-with-contracts-and-stdlib.zos.json');
