+++ packages/cli/src/test/scripts/set-admin.test.js
@@ -1,9 +1,10 @@
 'use strict';
-require('../setup');
+
+import { stubContractUpgradeable } from '../setup';
 
 import sinon from 'sinon';
 
-import { Contracts, Proxy, ProxyAdmin } from '@openzeppelin/upgrades';
+import { Proxy, ProxyAdmin } from '@openzeppelin/upgrades';
 import { accounts } from '@openzeppelin/test-environment';
 
 import push from '../../scripts/push';
@@ -11,9 +12,6 @@ import createProxy from '../../scripts/create';
 import setAdmin from '../../scripts/set-admin';
 import ProjectFile from '../../models/files/ProjectFile';
 import NetworkFile from '../../models/files/NetworkFile';
-
-import * as Compiler from '../../models/compiler/Compiler';
-import * as transpiler from '../../transpiler';
 
 const sandbox = sinon.createSandbox();
 
@@ -23,20 +21,7 @@ describe('set-admin script', function() {
   const network = 'test';
   const txParams = { from: owner };
 
-  beforeEach('stub getFromPathWithUpgradeable to simulate transpilation of contracts', async function() {
-    // stub getFromPathWithUpgradeable to fill upgradeable field with the same contract
-    sandbox.stub(Contracts, 'getFromPathWithUpgradeable').callsFake(function(targetPath, contractName) {
-      const contract = Contracts.getFromPathWithUpgradeable.wrappedMethod.apply(this, [targetPath, contractName]);
-      contract.upgradeable = contract;
-      return contract;
-    });
-    sandbox.stub(Compiler, 'compile');
-    sandbox.stub(transpiler, 'transpileAndSave');
-  });
-
-  afterEach(function() {
-    sandbox.restore();
-  });
+  stubContractUpgradeable(sandbox);
 
   const assertAdmin = async function(address, expectedAdmin, networkFile) {
     const actualAdmin = await Proxy.at(address).admin();
