+++ packages/cli/src/test/setup.js
@@ -3,6 +3,8 @@ import fs from 'fs-extra';
 process.env.NODE_ENV = 'test';
 
 import { ZWeb3, Contracts, Loggy, assertions } from '@openzeppelin/upgrades';
+import * as Compiler from '../models/compiler/Compiler';
+import * as transpiler from './../transpiler';
 import { provider } from '@openzeppelin/test-environment';
 
 import { OPEN_ZEPPELIN_FOLDER } from '../models/files/constants';
@@ -53,4 +55,21 @@ function setArtifactDefaults() {
 
   const defaults = process.env.SOLIDITY_COVERAGE ? DEFAULT_COVERAGE_TX_PARAMS : DEFAULT_TESTING_TX_PARAMS;
   Contracts.setArtifactsDefaults(defaults);
+}
+
+export function stubContractUpgradeable(sandbox) {
+  beforeEach('stub getFromPathWithUpgradeable to simulate transpilation of contracts', async function() {
+    // stub getFromPathWithUpgradeable to fill upgradeable field with the same contract
+    sandbox.stub(Contracts, 'getFromPathWithUpgradeable').callsFake(function(targetPath, contractName) {
+      const contract = Contracts.getFromPathWithUpgradeable.wrappedMethod.apply(this, [targetPath, contractName]);
+      contract.upgradeable = contract;
+      return contract;
+    });
+    sandbox.stub(Compiler, 'compile');
+    sandbox.stub(transpiler, 'transpileAndSave');
+  });
+
+  afterEach(function() {
+    sandbox.restore();
+  });
 }
